refactor: bug fixed with date picker component
