+++ src/components/appointmentsTable/table/style.tsx
